Move elasticsearch model classes into submodule
In preparation for introducing the ability to switch between model
classes, move the existing model classes in to a submodule.
