+++ h/api/models/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from h.api.models.elastic import Annotation, Document
+
+__all__ = ('Annotation', 'Document')
