+++ h/api/models/test/elastic_test.py
@@ -4,14 +4,13 @@ import itertools
 import re
 import urllib
 
-import jinja2
 import pytest
 from mock import patch
 from mock import PropertyMock
 
-from h.api import models
+from h.api.models.elastic import Annotation
 
-analysis = models.Annotation.__analysis__
+analysis = Annotation.__analysis__
 
 
 def test_strip_scheme_char_filter():
@@ -67,55 +66,55 @@ def groups(pattern, text):
 
 
 def test_uri():
-    assert models.Annotation(uri="http://foo.com").uri == "http://foo.com"
+    assert Annotation(uri="http://foo.com").uri == "http://foo.com"
 
 
 def test_uri_with_no_uri():
-    assert models.Annotation().uri == ""
+    assert Annotation().uri == ""
 
 
 def test_uri_when_uri_is_not_a_string():
     for uri in (True, None, 23, 23.7, {"foo": False}, [1, 2, 3]):
-        assert isinstance(models.Annotation(uri=uri).uri, unicode)
+        assert isinstance(Annotation(uri=uri).uri, unicode)
 
 
 def test_target_links_from_annotation():
-    annotation = models.Annotation(target=[{'source': 'target link'}])
+    annotation = Annotation(target=[{'source': 'target link'}])
     assert annotation.target_links == ['target link']
 
 
 def test_parent_returns_none_if_no_references():
-    annotation = models.Annotation()
+    annotation = Annotation()
     assert annotation.parent is None
 
 
 def test_parent_returns_none_if_empty_references():
-    annotation = models.Annotation(references=[])
+    annotation = Annotation(references=[])
     assert annotation.parent is None
 
 
 def test_parent_returns_none_if_references_not_list():
-    annotation = models.Annotation(references={'foo': 'bar'})
+    annotation = Annotation(references={'foo': 'bar'})
     assert annotation.parent is None
 
 
-@patch.object(models.Annotation, 'fetch', spec=True)
+@patch.object(Annotation, 'fetch', spec=True)
 def test_parent_fetches_thread_parent(fetch):
-    annotation = models.Annotation(references=['abc123', 'def456'])
+    annotation = Annotation(references=['abc123', 'def456'])
     annotation.parent
     fetch.assert_called_with('def456')
 
 
-@patch.object(models.Annotation, 'fetch', spec=True)
+@patch.object(Annotation, 'fetch', spec=True)
 def test_parent_returns_thread_parent(fetch):
-    annotation = models.Annotation(references=['abc123', 'def456'])
+    annotation = Annotation(references=['abc123', 'def456'])
     parent = annotation.parent
     assert parent == fetch.return_value
 
 
 @pytest.fixture
 def link_text(request):
-    patcher = patch('h.api.models.Annotation.link_text',
+    patcher = patch('h.api.models.elastic.Annotation.link_text',
                     new_callable=PropertyMock)
     request.addfinalizer(patcher.stop)
     return patcher.start()
@@ -123,14 +122,15 @@ def link_text(request):
 
 @pytest.fixture
 def title(request):
-    patcher = patch('h.api.models.Annotation.title',
+    patcher = patch('h.api.models.elastic.Annotation.title',
                     new_callable=PropertyMock)
     request.addfinalizer(patcher.stop)
     return patcher.start()
 
+
 @pytest.fixture
 def href(request):
-    patcher = patch('h.api.models.Annotation.href',
+    patcher = patch('h.api.models.elastic.Annotation.href',
                     new_callable=PropertyMock)
     request.addfinalizer(patcher.stop)
     return patcher.start()
@@ -138,9 +138,7 @@ def href(request):
 
 @pytest.fixture
 def hostname_or_filename(request):
-    patcher = patch('h.api.models.Annotation.hostname_or_filename',
+    patcher = patch('h.api.models.elastic.Annotation.hostname_or_filename',
                     new_callable=PropertyMock)
     request.addfinalizer(patcher.stop)
     return patcher.start()
-
-
